complete profile introduction text
