+++ lib/screens/complete_profile/components/complete_profile_form.dart
@@ -1,0 +1,201 @@
+import 'package:ecommerce/components/custom_surfix_icon.dart';
+import 'package:ecommerce/components/default_button.dart';
+import 'package:flutter/material.dart';
+
+import '../../../constants.dart';
+import '../../../size_config.dart';
+
+class CompleteProfileForm extends StatefulWidget {
+  @override
+  _CompleteProfileFormState createState() => _CompleteProfileFormState();
+}
+
+
+class _CompleteProfileFormState extends State<CompleteProfileForm> {
+  final _formKey = GlobalKey<FormState>();
+  final List<String> errors = [];
+  String firstName;
+  String lastName;
+  String phoneNumber;
+  String address;
+
+  void addError({String error}){
+    if(!errors.contains(errors))
+      setState(() {
+        errors.add(error);
+      });
+  }
+
+  void removeError({String error}){
+    if(errors.contains(error)){
+      setState(() {
+        errors.remove(error);
+      });
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Form(
+      key: _formKey,
+        child: Column(
+          children: [
+            buildFnameFormField(),
+            SizedBox(height: getProportionateScreenHeight(20),),
+            buildLastNameFormField(),
+            SizedBox(height: getProportionateScreenHeight(20),),
+            buildPhoneFormField(),
+            SizedBox(height: getProportionateScreenHeight(20),),
+            buildEmailFormField(),
+            SizedBox(height: getProportionateScreenHeight(20),),
+            DefaultButton(
+              text: "Continue",
+              press: (){},
+            ),
+
+
+          ],
+        ),
+
+
+
+    );
+  }
+
+  TextFormField buildLastNameFormField() {
+    return TextFormField(
+      onSaved:  (newValue) => lastName = newValue,
+      onChanged: (value) {
+        if (value.isNotEmpty){
+          removeError(error: kEmailNullError);
+        }else if(emailValidatorRegExp.hasMatch(value)){
+          removeError(error: kInvalidEmailError);
+        }
+        return null;
+      },
+      validator: (value){
+        if (value.isEmpty){
+          addError(error: kEmailNullError);
+        }else if(!emailValidatorRegExp.hasMatch(value)){
+          addError(error: kInvalidEmailError);
+          return "";
+        }
+        return null;
+      },
+      decoration:
+      InputDecoration(
+        labelText: "Last Name",
+        hintText: "Input your Last Name",
+        floatingLabelBehavior: FloatingLabelBehavior.always,
+        suffixIcon: CustomSuffixIcon(
+          svgIcon: "assets/icons/User.svg",
+        ),
+      ),
+    );
+  }
+
+  TextFormField buildFnameFormField() {
+    return TextFormField(
+      onSaved:  (newValue) => firstName = newValue,
+      onChanged: (value) {
+        if (value.isNotEmpty){
+          removeError(error: kEmailNullError);
+        }else if(emailValidatorRegExp.hasMatch(value)){
+          removeError(error: kInvalidEmailError);
+        }
+        return null;
+      },
+      validator: (value){
+        if (value.isEmpty){
+          addError(error: kEmailNullError);
+        }else if(!emailValidatorRegExp.hasMatch(value)){
+          addError(error: kInvalidEmailError);
+          return "";
+        }
+        return null;
+      },
+      decoration:
+      InputDecoration(
+        labelText: "First name",
+        hintText: "Input your First Name",
+        floatingLabelBehavior: FloatingLabelBehavior.always,
+        suffixIcon: CustomSuffixIcon(
+          svgIcon: "assets/icons/User.svg",
+        ),
+      ),
+    );
+  }
+
+  TextFormField buildPhoneFormField() {
+    return TextFormField(
+      keyboardType: TextInputType.phone,
+      onSaved:  (newValue) => phoneNumber = newValue,
+      onChanged: (value) {
+        if (value.isNotEmpty) {
+          removeError(error: kPhoneNumberNullError);
+        }
+        return null;
+      },
+      validator: (value){
+        if (value.isEmpty){
+          addError(error: kPhoneNumberNullError);
+          return "";
+        }
+        return null;
+      },
+      decoration:
+      InputDecoration(
+        labelText: "Phone Number",
+        hintText: "Input your Phone number",
+        floatingLabelBehavior: FloatingLabelBehavior.always,
+        suffixIcon: CustomSuffixIcon(
+          svgIcon: "assets/icons/Phone.svg",
+        ),
+      ),
+    );
+  }
+
+  TextFormField buildEmailFormField() {
+    return TextFormField(
+      keyboardType: TextInputType.emailAddress,
+      onChanged: (value) {
+        if (value.isNotEmpty && errors.contains(kEmailNullError)){
+          setState(() {
+            errors.remove(kEmailNullError);
+          });
+        }else if(emailValidatorRegExp.hasMatch(value) &&
+            errors.contains(kEmailNullError)){
+          setState(() {
+            errors.remove(kEmailNullError);
+          });
+
+        }
+        return null;
+      },
+      validator: (value){
+        if (value.isEmpty && !errors.contains(kEmailNullError)){
+          setState(() {
+            errors.add(kEmailNullError);
+          });
+        }else if(emailValidatorRegExp.hasMatch(value) &&
+            !errors.contains(kEmailNullError)){
+          setState(() {
+            errors.add(kEmailNullError);
+
+          });
+        }
+        return null;
+      },
+      decoration:
+      InputDecoration(
+        labelText: "Address",
+        hintText: "Input your Address",
+        floatingLabelBehavior: FloatingLabelBehavior.always,
+        suffixIcon: CustomSuffixIcon(
+          svgIcon: "assets/icons/Location point.svg",
+        ),
+      ),
+    );
+  }
+
+}
